refactor(chat): restore vanilla framework layout using ChatWidget in ChatApp
